Fix failing PNG write from ARGB input
Summary:
@public

The logic of `PNG_COLOR_TYPE_RGBA | PNG_FORMAT_FLAG_AFIRST` was not sound. A format flag is not of the same (semantic) type as a color type.

In this case the `png_set_IHDR(...)` method would fail with "Invalid IHDR data". Instead we need to call the method to swap alpha channels separately.

Actually detected when landing instrumentation tests (see next diff/commit). Let's have this as a lecture on the value of testing. Do tests kids!

Reviewed By: AurelC2G

Differential Revision: D15413845

fbshipit-source-id: a35ab0aa155fa71aa225a3ac2a2e69a91d9b5f20