Rename VectorImageSink to VectorBitmapImageSink
Summary:
For adding Webassembly interface for Spectrum we need to separate VectorImageSink to VectorEncodeImageSink and VectorBitmapImageSink, as it done for VectorImageSource. This commit is first part of this separation: I've renamed VectorImageSink -> VectorBitmapImageSink.

After landing of this changes, I'm going to create new diff with adding VectorEncodedImageSink class and new unit tests.

@public
Rename VectorImageSink to VectorBitmapImageSink

Reviewed By: AurelC2G

Differential Revision: D16163200

fbshipit-source-id: 8d61c296f4e67738ead4b85ba179d4dc29093876