Add SpectrumPluginPlatform
Summary:
@public

This diff adds a new plugin which exposes the platform decoder on Android (i.e. `BitmapFactory`) as decompressors for various image formats to Spectrum. The benefit is that one e.g. no longer needs to ship the WebP plugin if you are just interested in decoding WebP files. Also it newly adds support for GIF and HEIC files on Android versions that support it. However, note that we are crossing the JNI barrier a bit more often and allocate an additional (temporary) Bitmap which can be a bottleneck.

Before explaining the underlying structure, be warned. The changes are a bit intricate and cover: Java life cycle, JNI calls in both directions, reflection, C++ iterators, RAII bitmap locks, and more. Hope you enjoy it.

How does this work:

 - The `JniSpectrumPluginPlatform` creates a new Plugin which references the `JniPlatformDecompressor` for various file formats – i.e. it creates multiple `DecompressorFactory` instances. The `JniPlatformDecompressor` implements our usual `IDecompressor` interface but is agnostic to the incoming file format.
 - When being instantiated, it will create an instance of the `SpectrumPlatformDecompressor` in the Java world and pass the entire `IImageSource` content as a `byte[]` array
 - When the interface methods are called, it will lazily request the image specification or Bitmap data and hold a reference locally for subsequent calls.
   - The `image::Specification` is derived from the `Bitmap#Options` object in Java
   - The `image::Scanline` is copied out of the temporary locked `JBitmap` object using a RAII helper
 - Tests are using our normal declarative testing framework

Missing bits:

 - Test with HEIC files (probably not to be included as automated tests as only supported on physical devices)
 - Support for sampling (at least the JPEG 1/2, 1/4, 1/8 ones)
 - Optimization: using the `RegionDecoder` to have less allocations

Reviewed By: cuva

Differential Revision: D15758468

fbshipit-source-id: c1ae9f3cdd8dba0943ee1dcdc638201f95beb2f6