feat(website): add DocSearch as recommended by docusaurus (#207)
Summary:
👋 team,

I am working on DocSearch. [We have an integration for Docusaurus](https://docusaurus.io/docs/en/search#enabling-the-search-bar). We thought it is a shame you can not also used this search that you can [find on reactjs for example](https://reactjs.org/).

This PR will add DocSearch to the documentation website. It will allow an user to have a learn-as-you-type experience by displaying results thanks to a dropdown in a live way.

Let me know if you need anything.
Pull Request resolved: https://github.com/facebookincubator/spectrum/pull/207

Reviewed By: AurelC2G

Differential Revision: D21276463

Pulled By: diegosanchezr

fbshipit-source-id: d94ebceaa84ff1845152ca94eadfd058d0db1d33