Avoid double decoding for AVIF
Summary:
In order to get the specifications of an AVIF image, `AvifDecompressor` was decoding the full image. Because we don't want to decode it again later on to grab the decoded bitmap, we cache the decoded result within the decompressor instance.

When computing which rules to apply for a transcode (in `spectrum::core::OperationBuilder::build`), Spectrum creates an `AvifDecompressor` instance to get the specs of the input AVIF image.
Later on, when actually running the transcode, `spectrum::Spectrum::_run` creates a different `AvifDecompressor` instance in order to actually decode the image.

Because these two instances are different, they do not share the cached bitmap, and therefore we end up decoding the image twice.

This diff changes the way we decode the AVIF image, so that when we just need the specs and not the full bitmap, we only parse the container to read the dimensions, and do not decode the image.

Reviewed By: erikandre, defHLT

Differential Revision: D24109313

fbshipit-source-id: fb65fcd457a7ba5e7f4f76541abafda741d2a068