Add AllocatedImageSink to save on memory copying
Summary:
@public
Adding new AllocatedImageSink that allows to pass an allocated buffer in order to avoid copying bytes when using VectorImageSink.

It is mostly useful for bitmaps as we can easily predict the buffer size required depending on required image dimensions and the memory savings should be more significant. We're not exposing a sink for encoded image as the encoded file size might be harder to predict but it can be added if needed.

Reviewed By: AurelC2G

Differential Revision: D25193923

fbshipit-source-id: b05f4d0c4e8900a093e117ea9d506369ef2de907