Adding Social Banner in Support of Ukraine (#1839)
Summary:
Our mission at [Meta Open Source](https://opensource.facebook.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

<img width="1680" alt="image" src="https://user-images.githubusercontent.com/77028195/156680404-acfcc446-3e82-4dfc-901f-c263b5411840.png">

Pull Request resolved: https://github.com/facebookincubator/spectrum/pull/1839

Reviewed By: cuva

Differential Revision: D34631615

Pulled By: zmroczek

fbshipit-source-id: 2aae98078c78ee3ab6cb7e1a2ee7bb83defa94d3